update and minor improvement

+++ projects/printf/include/type_conversion.h
@@ -1,7 +1,7 @@
 #ifndef __TYPE_CONV_H__
 #define __TYPE_CONV_H__
 
-void _itob(int n, char *s, int b);
-void _ftoa(double n, char *s);
+void _itob(long long n, char *s, int b);
+void _ftoa(long double n, char *s);
 
 #endif
